Added a check to ensure the correct dependencies are called.

+++ qr_decomposition/1_systolic_array/compile_hardware.sh
@@ -4,6 +4,12 @@ set -e
 
 current_date_time="`date +%Y-%m-%dT%H:%M:%S`";
 
+if [ ! -d "streamblocks-examples" ]; then
+  echo "Missing the streamblocks-examples repository in this directory. It is required to call the sqrt() function."
+  echo "Pull with 'git clone https://github.com/streamblocks/streamblocks-examples.git'"
+  exit 1
+fi
+
 rm -rf myproject
 
 streamblocks vivado-hls --set experimental-network-elaboration=on --set bypass-AM-generation=on --source-path qrd_systolic.cal:streamblocks-examples/system --target-path myproject qrd.Top
